Dante-related fixes

+++ coppafish/pipeline/extract_run.py
@@ -16,6 +16,7 @@ def run_extract(
     nbp_file: NotebookPage,
     nbp_basic: NotebookPage,
     nbp_scale: NotebookPage,
+    return_image_t_raw: bool = False,
 ) -> Tuple[NotebookPage, NotebookPage, Optional[np.ndarray]]:
     """
     This reads in images from the raw `nd2` files, filters them and then saves them as `config[extract][file_type]`
@@ -40,11 +41,12 @@ def run_extract(
         - See `'extract'` and `'extract_debug'` sections of `notebook_comments.json` file for description of the
             variables in each page.
     """
-    # TODO: Refactor this function to make it more readable
     # initialise notebook pages
     if not nbp_basic.is_3d:
         # config["deconvolve"] = False  # only deconvolve if 3d pipeline
         raise NotImplementedError(f"coppafish 2d is not in a stable state, please contact a dev to add this. Sorry! ;(")
+    if return_image_t_raw:
+        assert len(nbp_basic.use_tiles) == 1, "The notebook must contain a single tile to return its image"
 
     start_time = time.time()
     nbp = NotebookPage("extract")
@@ -61,32 +63,24 @@ def run_extract(
         # always have anchor as first round after imaging rounds
         round_files = nbp_file.round + [nbp_file.anchor]
         use_rounds = np.arange(len(round_files))
-        n_images = (
-            (len(use_rounds) - 1) * len(nbp_basic.use_tiles) * len(nbp_basic.use_channels)
-            + len(nbp_basic.use_tiles) * len(use_channels_anchor)
-            + len(nbp_basic.use_tiles) * len(nbp_basic.use_rounds) * config["continuous_dapi"]
-        )
     else:
         round_files = nbp_file.round
         use_rounds = nbp_basic.use_rounds
-        n_images = len(use_rounds) * len(nbp_basic.use_tiles) * len(nbp_basic.use_channels)
 
     # If we have a pre-sequencing round, add this to round_files at the end
     if nbp_basic.use_preseq:
         round_files = round_files + [nbp_file.pre_seq]
         use_rounds = np.arange(len(round_files))
         pre_seq_round = len(round_files) - 1
-        n_images += len(nbp_basic.use_tiles) * len(nbp_basic.use_channels)
     else:
         pre_seq_round = None
 
-    return_image_t = len(nbp_basic.use_tiles) == 1
-    if return_image_t:
+    if return_image_t_raw:
         image_t = np.zeros(
             (
-                max(use_rounds) + 1,
-                max(nbp_basic.use_channels + [nbp_basic.dapi_channel] * config["continuous_dapi"]) + 1,
-                max(nbp_basic.use_z) + 1,
+                nbp_basic.n_rounds + nbp_basic.n_extra_rounds,
+                nbp_basic.n_channels,
+                nbp_basic.nz,
                 nbp_basic.tile_sz,
                 nbp_basic.tile_sz,
             ),
@@ -94,22 +88,41 @@ def run_extract(
         )
     else:
         image_t = None
+    nbp_debug.pixel_unique_values = np.full(
+        (
+            nbp_basic.n_tiles,
+            nbp_basic.n_rounds + nbp_basic.n_extra_rounds,
+            nbp_basic.n_channels,
+            np.iinfo(np.uint16).max,
+        ),
+        fill_value=0,
+        dtype=int,
+    )
+    nbp_debug.pixel_unique_counts = nbp_debug.pixel_unique_values.copy()
 
     with tqdm(
-        total=n_images, desc=f"Loading raw {nbp_file.raw_extension} tiles and saving as {config['file_type']}"
+        total=(len(use_rounds) - 1)
+            * len(nbp_basic.use_tiles)
+            * (len(nbp_basic.use_channels) + 1 if nbp_basic.dapi_channel is not None else 0)
+            + len(nbp_basic.use_tiles) * len(use_channels_anchor),
+        desc=f"Extracting raw {nbp_file.raw_extension} files to {config['file_type']}",
     ) as pbar:
         for r in use_rounds:
             round_dask_array, metadata = utils.raw.load_dask(nbp_file, nbp_basic, r=r)
 
-            if metadata is not None:
-                # Dump all metadata to pickle file
-                with open(os.path.join(nbp_file.tile_unfiltered_dir, f"nd2_metadata_r{r}.pkl"), "w") as file:
-                    pickle.dump(metadata, file)
+            metadata_path = os.path.join(nbp_file.tile_unfiltered_dir, f"nd2_metadata_r{r}.pkl")
+            if not os.path.isfile(metadata_path):
+                if metadata is not None:
+                    # Dump all metadata to pickle file
+                    with open(metadata_path, "wb") as file:
+                        pickle.dump(metadata, file)
 
             if r == nbp_basic.anchor_round:
                 use_channels = use_channels_anchor
             else:
-                use_channels = nbp_basic.use_channels + [nbp_basic.dapi_channel] * config["continuous_dapi"]
+                use_channels = nbp_basic.use_channels.copy()
+                if nbp_basic.dapi_channel is not None:
+                    use_channels += [nbp_basic.dapi_channel]
 
             # convolve_2d each image
             for t in nbp_basic.use_tiles:
@@ -137,16 +150,23 @@ def run_extract(
                             file_exists = tiles_io.image_exists(file_path, config["file_type"])
                         pbar.set_postfix({"round": r, "tile": t, "channel": c, "exists": str(file_exists)})
 
-                        # TODO: Save the raw uint16 dask array file
-                        im = utils.raw.load_image(nbp_file, nbp_basic, t, c, round_dask_array, r, nbp_basic.use_z)
-                        im = im.astype(np.uint16, casting="safe")
-                        # yxz -> zyx
-                        im = im.transpose((2, 0, 1))
-                        tiles_io._save_image(im, file_path, config["file_type"])
-                        if return_image_t:
+                        if file_exists:
+                            im = tiles_io._load_image(file_path, config["file_type"])
+                        if not file_exists:
+                            im = utils.raw.load_image(nbp_file, nbp_basic, t, c, round_dask_array, r, nbp_basic.use_z)
+                            im = im.astype(np.uint16, casting="safe")
+                            assert np.sum(im) != 0, f"Extracted image {t=}, {r=}, {c=} contains all zeros"
+                            # yxz -> zyx
+                            im = im.transpose((2, 0, 1))
+                            tiles_io._save_image(im, file_path, config["file_type"])
+                        if return_image_t_raw:
                             image_t[r, c] = im
+                        pixel_unique_values, pixel_unique_counts = np.unique(im, return_counts=True)
                         del im
-                        pbar.update(1)
+                        nbp_debug.pixel_unique_values[t][r][c][: pixel_unique_values.size] = pixel_unique_values
+                        nbp_debug.pixel_unique_counts[t][r][c][: pixel_unique_counts.size] = pixel_unique_counts
+                    pbar.update(1)
+            del round_dask_array
     end_time = time.time()
     nbp_debug.time_taken = end_time - start_time
     return nbp, nbp_debug, image_t
